refactor(acceptance): extract PantherAcceptanceTestCase and seeding trait
Consolidate duplication observed across the 5 shipped Small-tier
Acceptance E2E tests into reusable Support/ infrastructure so
Medium-tier ports (Bb/Cc/Dd/Ee/Ff/SvcCode) consume from day one
instead of re-duplicating.

Hybrid shape: abstract base class for near-universal infrastructure
(client lifecycle, admin login, Knockout-ready gate, Product
Registration modal dismissal, iframe switch, tab-active assertion,
XPath/JS constants); opt-in trait for UI-driven patient + encounter
seeding (only Medium-tier ports need this).

New files under tests/Acceptance/Support/:
 - PantherAcceptanceTestCase.php -- abstract base extending
   PHPUnit TestCase. Provides tearDown Chrome-quit discipline,
   requireClient() nullable-narrow, performLoginAsAdmin() with
   three-gate wait chain and diagnostic self::fail on each gate,
   dismissProductRegistrationModalIfPresent() with split try/catch
   scope, switchToIFrame(), waitAndClick(), assertActiveTab(),
   and constants for login/shell titles, XPaths, JS gates.
 - UiSeedingTrait.php -- addPatientViaUi(), addEncounterViaUi(),
   setInputValue(), setSelectValue(), plus fixture-identity
   constants (Ftest Ltest, DOB, sex, category id, reason).

Refactored 5 existing tests to consume the new Support/:
 - E2eCriticalPathTest      -- extends base, uses performLoginAsAdmin
 - AaLoginAcceptanceTest    -- extends base (auth-negative flows;
   uses lifecycle + constants but not the login helper)
 - GgUserMenuLinksAcceptanceTest -- extends base, uses login helper
   and assertActiveTab + waitAndClick
 - FrontPaymentCssContrastAcceptanceTest -- extends base for
   lifecycle; keeps its scoped-down login inline because the direct
   GET flow doesn't need the Knockout-ready gate or modal dismiss
 - KkEncounterFormNavbarUrlAcceptanceTest -- extends base and
   `use UiSeedingTrait`, deletes inline seeding + iframe helpers

Preserved exactly: every assertion, group attribute, test method
name, and every WHY comment (release-image XPath gotchas, race-
condition anti-flake reasoning, XPath calibration notes).

Local verification against booted release image on port 8580:
5 test classes, 10 test methods, 93 assertions, all passing in
~49s (identical to baseline). PHPStan level 10 clean, phpcs
clean (pre-existing sqlconf.php warnings are unrelated), rector
clean for extracted files.

LOC delta on the 5 test files: 1647 to 808 (-839, -50.9 percent);
Support/ additions total 682, net -157 across tests + support.
