+++ tests/Acceptance/E2eCriticalPathTest.php
@@ -12,13 +12,10 @@ declare(strict_types=1);
 
 namespace OpenEMR\Tests\Acceptance;
 
-use Facebook\WebDriver\Exception\TimeoutException;
-use Facebook\WebDriver\JavaScriptExecutor;
 use Facebook\WebDriver\WebDriverBy;
 use OpenEMR\Tests\Acceptance\Support\BrowserSession;
+use OpenEMR\Tests\Acceptance\Support\PantherAcceptanceTestCase;
 use PHPUnit\Framework\Attributes\Group;
-use PHPUnit\Framework\TestCase;
-use Symfony\Component\Panther\Client;
 
 /**
  * Browser-driven end-to-end acceptance test.
@@ -44,97 +41,19 @@ use Symfony\Component\Panther\Client;
  */
 #[Group('fresh-install')]
 #[Group('post-upgrade')]
-final class E2eCriticalPathTest extends TestCase
+final class E2eCriticalPathTest extends PantherAcceptanceTestCase
 {
-    private ?Client $client = null;
-
-    protected function tearDown(): void
-    {
-        if ($this->client !== null) {
-            // Panther leaves a Chrome subprocess + ChromeDriver session
-            // dangling if not explicitly quit. Local runs accumulate
-            // zombies; CI runners recycle so it doesn't matter there,
-            // but tearDown discipline keeps local dev clean.
-            $this->client->quit();
-            $this->client = null;
-        }
-    }
-
     public function testAdminCanLogInAndKnockoutMainMenuRenders(): void
     {
         $this->client = BrowserSession::create();
-        // Panther's request() takes either an absolute URL or a path
-        // relative to `external_base_uri` (set to ACCEPTANCE_ARTIFACT_URL
-        // in the BrowserSession factory).
-        $this->client->request('GET', '/interface/login/login.php?site=default');
-
-        self::assertSame(
-            'OpenEMR Login',
-            $this->client->getTitle(),
-            'Login page title should be "OpenEMR Login" — a different title means the login route rendered something unexpected (setup wizard, error page, blank shell)',
-        );
-
-        // Fill and submit the login form. openemr's login form
-        // requires two hidden fields (languageChoice + new_login_session_management)
-        // in addition to the visible authUser/clearPass — those are
-        // pre-populated by the form template, so submitting the
-        // form element (rather than reconstructing a POST) picks
-        // them up automatically.
-        $this->client->submitForm('login-button', [
-            'authUser' => 'admin',
-            'clearPass' => 'pass',
-        ]);
-
-        // Post-login lands on /interface/main/tabs/main.php, which
-        // loads the Knockout-templated main menu asynchronously.
-        // waitForVisibility on #mainMenu confirms the DOM node
-        // exists; the executeScript check below confirms Knockout
-        // actually populated it with children (i.e. bindings
-        // applied, not just an empty shell).
-        try {
-            $this->client->waitFor('#mainMenu', 30);
-        } catch (TimeoutException) {
-            self::fail(
-                'Post-login #mainMenu element never appeared (30s timeout). '
-                . 'Landing URL: ' . $this->client->getCurrentURL() . '. '
-                . 'Title: ' . $this->client->getTitle() . '. '
-                . 'This likely means the login POST did not authenticate, or '
-                . 'the post-login shell (interface/main/tabs/main.php) never rendered.'
-            );
-        }
-
-        // Knockout-ready gate: bindings have applied, the menu template
-        // ran, so #mainMenu has child nodes. Matches the source-side
-        // BaseTrait's waitForAppReady contract. Without this check we
-        // could pass on a broken JS build that renders the SPA shell
-        // but leaves the menu empty.
-        //
-        // Type-hint the closure's $driver as JavaScriptExecutor so
-        // phpstan can resolve executeScript() (the base WebDriver
-        // interface types the callable arg as `mixed`; source-side
-        // E2E tests baseline this on principle, but we prefer to
-        // type-narrow at the source instead).
-        try {
-            $this->client->wait(30)->until(fn(JavaScriptExecutor $driver): bool => (bool) $driver->executeScript(
-                'return document.getElementById("mainMenu")?.children.length > 0'
-            ));
-        } catch (TimeoutException) {
-            $rawDiagnostics = $this->client->executeScript(<<<'JS_WRAP'
-                return JSON.stringify({
-                    url: location.href,
-                    title: document.title,
-                    knockoutLoaded: typeof ko !== "undefined",
-                    mainMenuChildCount: document.getElementById("mainMenu")?.children.length ?? 0,
-                });
-            JS_WRAP);
-            $diagnostics = is_string($rawDiagnostics) ? $rawDiagnostics : 'unavailable';
-            self::fail(
-                'Knockout main menu never populated (30s timeout after #mainMenu appeared). '
-                . "Page state: {$diagnostics}. This means the SPA shell loaded but the "
-                . 'JS framework (Knockout) either did not load or did not apply bindings — '
-                . 'symptom of a broken webpack build, missing JS asset, or template regression.'
-            );
-        }
+        // performLoginAsAdmin encapsulates the full login + Knockout-
+        // ready gate discipline (title change → #mainMenu appears →
+        // Knockout populates children) with diagnostic self::fail
+        // messages on each timeout. That's the exact scope this test
+        // asserts on — the post-login-shell-boots-cleanly signal —
+        // so the assertion here just checks the visible menu content
+        // after the base-class gate lands.
+        $this->performLoginAsAdmin();
 
         // At this point the app is fully booted — assert one visible
         // menu label to prove the menu isn't just full of empty nodes.
@@ -143,7 +62,8 @@ final class E2eCriticalPathTest extends TestCase
         // every skin). If this ever fails, the failure message points
         // at the specific expectation being wrong rather than a vague
         // "menu is broken."
-        $calendarMenu = $this->client->findElements(
+        $client = $this->requireClient();
+        $calendarMenu = $client->findElements(
             WebDriverBy::xpath('//div[@id="mainMenu"]//div[text()="Calendar"]'),
         );
         self::assertNotEmpty(
